SPM-1166: add release_versions into /advisories and /system_advisories
